improve Call[WithContext] expansion for the generated enum type
Instead of using `IntoOpt` and making use of the Call[WithContext] impl of the converted-to generated `opt` type the macro now generates a match based impl for the `enum` type own its own. This should be more intuitive for readability as well as more idiomatic Rust code and likely also more efficient.