add initial benchmark impl
So far the generated <enum> type highly outperforms the <opt> type which is sad. However, this may be due to the fact that Rust/LLVM can optimize the loop+match dispatch of the <enum> type in a way that distributes the jumps which has a similar effect to computed gotos which naturally outperform the indirect jumps from the <opt> type implementation. If this is true the big factor for the performance difference is that the branch predictor has a much simpler life predicting the computed-goto compared to the indirect calls of the <opt> type. Nevertheless, we may improve performance of the <opt> case by further improving performance of the underlying interpreter to make the Rust/LLVM backend generate more tail calls. I guestimate that panic infrastructure prevents Rust/LLVM from ideally optimizing the <opt> impl.